+++ CHANGELOG.md
@@ -14,6 +14,8 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 - `cluster instance deploy` now creates a new cluster config context upon successful cluster creation
 - `housekeeping realms create` has completely different (and incompatible) semantics: it now allows supplying
   either a public, private or no key, and will create a new config context accordingly
+- `cluster {install,upgrade,uninstall}-operator` commands are deprecated, print instructions to
+  perform the same tasks using Helm.
 
 ## [0.11.2] - 2020-10-22
 ### Added